move back to read to allow multiple sending of requests
## Description
The issue with using scanner is it waits for an EOF err to break. The
problem is until the connection is closed the server will not return an
EOF response ergo the scanner will still keep running thus blocking
subsequent requests from being made.
Need to find other ways to handle it, so far what I can do is to read
the length of the header for the content-length and use that to create
my buffer size or create a maxBytes so I can read and count to that
number. Might go with maxBytes because I can stream the data as I read
to the buffer without having to keep a lot of data in memory.